fix: moved PersonPanel to its own dir
