+++ src/dataDisplay/index.js
@@ -8,8 +8,8 @@ export { default as Icon } from './icon';
 export { default as Image } from './image';
 export { default as List } from './list';
 export { default as PersonPanel } from './personPanel';
-export { default as PersonPanelDetails } from './personPanelDetails';
-export { default as PersonPanelSummary } from './personPanelSummary';
+export { default as PersonPanelDetails } from './personPanel/personPanelDetails';
+export { default as PersonPanelSummary } from './personPanel/personPanelSummary';
 export { default as TelephoneLink } from './telephoneLink';
 export { default as Table } from './table';
 export { default as Tooltip } from './tooltip';
