+++ src/dataDisplay/personPanel/personPanelConstants.js
